+++ CHANGELOG.md
@@ -16,10 +16,14 @@
   `measurements <https://docs.pennylane.ai/en/stable/introduction/measurements.html>`_ themselves.
   [(#405)](https://github.com/PennyLaneAI/pennylane-qiskit/pull/405)
 
-* Added support for coverting conditional operations based on mid-circuit measurements and
+* Added the support for converting conditional operations based on mid-circuit measurements and
   two of the ``ControlFlowOp`` operations - ``IfElseOp`` and ``SwitchCaseOp`` when converting
   a ``QuantumCircuit`` using `load`.
   [(#417)](https://github.com/PennyLaneAI/pennylane-qiskit/pull/417)
+
+* Qiskit's classical ``Expr`` conditionals can also be used with the supported
+  ``ControlFlowOp`` operations.
+  [(#432)](https://github.com/PennyLaneAI/pennylane-qiskit/pull/432)
 
 * Added conversion support for more Qiskit gates to native PennyLane operations -
   ``Barrier``, ``CYGate``, ``CHGate``, ``CPhase``, ``CCZGate``, ``ECRGate``, and ``GlobalPhaseGate``.